refactor: 파일 구조 리팩터링, 결과 페이지 스타일 리팩터링
