+++ src/pages/RandomPick/RandomSelect/ConditionBox/style.ts
