+++ src/pages/RandomPick/RandomSelect/RadioBox/index.tsx
